Update service name in Jenkins Environment Variables
Corrected a typo in the Jenkinsfile, specifically in the SERVICE_NAME environment variable. The service name was previously "csctracker_services_metric" and has been updated to "csctracker_services_metrics". This change is to ensure consistency in naming across all our services and facilitates automated tasks such as deployment and monitoring.